[tests-only][full-ci] Backport test for branded client (#12281)
* [tests-only][full-ci] automate branded client (#12175)

add test codes to automate tests for branded client

* [tests-only][full-ci] update sync related test for automating branded client (#12230)

automante branded client - update test for syncing

* get password of predefined users from created users

+++ test/gui/shared/scripts/helpers/api/provisioning.py
@@ -32,6 +32,16 @@ def add_user_to_group(user, group_name):
 
 
 def create_user(username):
+    if get_config('client_name') != 'ownCloud':
+        user = UserHelper.test_users[username]
+        created_users[username] = {
+            'username': user.username,
+            'password': user.password,
+            'displayname': user.displayname,
+            'email': user.email,
+        }
+        return
+
     if username in UserHelper.test_users:
         user = UserHelper.test_users[username]
     else:
@@ -61,8 +71,9 @@ def create_user(username):
 
 def delete_created_users():
     for username, user_info in list(created_users.items()):
-        if get_config('ocis'):
-            ocis.delete_user(user_info['id'])
-        else:
-            oc.delete_user(user_info['id'])
+        if get_config('client_name') == 'ownCloud':
+            if get_config('ocis'):
+                ocis.delete_user(user_info['id'])
+            else:
+                oc.delete_user(user_info['id'])
         del created_users[username]
